+++ package.json
@@ -5,6 +5,9 @@
   "bin": {
     "mcsc": "bin/mcsc"
   },
+  "scripts": {
+    "build": "npx tsc"
+  },
   "publishConfig": {
     "access": "public",
     "@piarre:registry": "https://registry.npmjs.org/"
